Updated scripts
Added a description for evaluations_functions.py

Cleaned experiment_config.py
As of now, still lacking a global time variable, but this is easy to implement and I will do so in further experiments.

Changed make_datasets.py to generate a few smaller datasets to run tests on and save to a new directory.