Added the relevant plots and removed the irrelevant ones.
These are the plots obtained with the newly generated results.

There's also a plot illustrating the dataset distribution and a plot which shows the difference between the same experiment ran twice on peregrine. The latter illustrates how much the runtime is affected by other factors, which does not appear to be too bad.