Updating json files
Renamed json files to be more in line with experiment.

Also removed the dataset with 10_000 objects, since it was not described in the synthetic dataset parameters.

The results of running the methods on the real world datasets are also included. Note that one result from the UDFS method is still missing.